requirements.txt
attrs==19.3.0
Automat==20.2.0
cffi==1.14.0
configparser==5.0.0
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
lxml==4.5.0
parsel==1.6.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyMySQL==0.9.3
pyOpenSSL==19.1.0
python-dotenv==0.13.0
queuelib==1.5.0
Scrapy==2.1.0
scrapy-selenium==0.0.7
selenium==3.141.0
service-identity==18.1.0
six==1.14.0
SQLAlchemy==1.3.17
Twisted==20.3.0
Unidecode==1.1.1
urllib3==1.25.9
w3lib==1.22.0
zope.interface==5.1.0
